chore: updated unittests
